backend/requirements_simple.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-telegram-bot==20.7
python-dotenv==1.0.0